Refactoring + multiple fixes
