+++ rover_gui/CMakeLists.txt
@@ -22,6 +22,7 @@ set(CMAKE_AUTOMOC ON)
 set(CMAKE_AUTORCC ON) 
   
 include_directories(
+  src
   ui/include
   ${rover_msgs_INCLUDE_DIRS}
   ${LIBSSH_INCLUDE_DIR}
@@ -46,6 +47,10 @@ add_executable(main_gui src/main_gui.cpp
                         src/Global/QTmpFolderManager.hpp
                         src/Global/QTmpFolderManager.cpp
 
+                        src/QSshFileExplorer/Worker/LibSshSupportModule.hpp
+                        src/QSshFileExplorer/Worker/LibSshSupportModule.cpp
+                        src/QSshFileExplorer/Worker/QSshWorker.hpp
+                        src/QSshFileExplorer/Worker/QSshWorker.cpp
                         src/QSshFileExplorer/QDownloadedFileManager.hpp
                         src/QSshFileExplorer/QDownloadedFileManager.cpp
                         src/QSshFileExplorer/QFileItem.hpp
@@ -56,8 +61,6 @@ add_executable(main_gui src/main_gui.cpp
                         src/QSshFileExplorer/QIconManager.cpp
                         src/QSshFileExplorer/QSshFileExplorerWidget.hpp
                         src/QSshFileExplorer/QSshFileExplorerWidget.cpp
-                        src/QSshFileExplorer/QSshWorker.hpp
-                        src/QSshFileExplorer/QSshWorker.cpp
                         src/QSshFileExplorer/QTreeViewExplorer.hpp
                         src/QSshFileExplorer/QTreeViewExplorer.cpp)
 add_dependencies(main_gui run_generate_ui)
